Move all components into separate folder
