+++ src/components/Alert.js
@@ -9,7 +9,7 @@ import { withStyles } from "@material-ui/core/styles";
 import PropTypes from "prop-types";
 import React from "react";
 
-import { t } from "./utils";
+import { t } from "../utils";
 
 const styles = theme => ({
   root: {},
