+++ src/components/Link.js
@@ -3,8 +3,8 @@ import MuiLink from "@material-ui/core/Link";
 import PropTypes from "prop-types";
 import React, { Children } from "react";
 
-import AdminContext from "./context/AdminContext";
-import { toQuery } from "./utils";
+import AdminContext from "../context/AdminContext";
+import { toQuery } from "../utils";
 
 export default class Link extends React.Component {
   static contextType = AdminContext;
