+++ src/pages/LoginPage.js
@@ -5,7 +5,7 @@ import Typography from "@material-ui/core/Typography";
 import PropTypes from "prop-types";
 import React from "react";
 
-import Logo from "../Logo";
+import Logo from "../components/Logo";
 import DefaultLoginForm from "./LoginPageForm";
 
 const DialogTitle = withStyles(theme => ({
